% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpR2_Function.R
\name{cmp.R2}
\alias{cmp.R2}
\title{Compute R2 with a specified C matrix}
\usage{
cmp.R2(c, x, SigHat, beta, NS.adj = F)
}
\arguments{
\item{c}{Contrast matrix for fixed effects}

\item{x}{Fixed effects design matrix}

\item{SigHat}{estimated model covariance matrix}

\item{beta}{fixed effects estimates}

\item{NS.adj}{if True, the denominator degrees of freedom are
adjusted to replicate the R squared proposed by Nakagawa and Schielzeth.}
}
\value{
A vector with the Wald statistic (ncp), approximate Wald F
statistic (F), numerator degrees of freedom (v1), denominator degrees
of freedom (v2), and the specified r squared value (Rsq)
}
\description{
Compute R2 with a specified C matrix
}
\examples{
library(nlme)
library(lme4)
lmemod = lme(distance ~ age*Sex, random = ~1|Subject, data = Orthodont)
X = model.matrix(lmemod, data = Orthodont)
SigHat = extract.lme.cov(lmemod, data = Orthodont)
beta = fixef(lmemod)
p = length(beta)

C = cbind(rep(0, p-1),diag(p-1))
partial.c = make.partial.C(p-1,p,2)

cmp.R2(c=C, x=X, SigHat=SigHat, beta=beta)
cmp.R2(c=partial.c, x=X, SigHat=SigHat, beta=beta)
}

