% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc_Function.R
\name{make.partial.C}
\alias{make.partial.C}
\title{Generate partial contrast matrices}
\usage{
make.partial.C(rows, cols, index)
}
\arguments{
\item{rows}{Number of rows in the contrast matrix}

\item{cols}{Number of columns in the contrast matrix}

\item{index}{A number corresponding to the position of the fixed effect
in the vector of fixed effect parameter estimates.}
}
\value{
A contrast matrix designed to test the fixed effect
 corresponding to \code{index} in the vector of fixed effects.
}
\description{
Generate partial contrast matrices
}
\examples{
make.partial.C(4, 5, 2)
make.partial.C(4, 5, 3)
make.partial.C(4, 5, 2:4)
}
