% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r2_Function.R
\name{print.R2}
\alias{print.R2}
\title{Print the contents of an R2 object}
\usage{
\method{print}{R2}(x, ...)
}
\arguments{
\item{x}{an object of class R2}

\item{...}{other arguments passed to the print function.}
}
\description{
Print the contents of an R2 object
}
