% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_sgv_est_Function.R
\name{r2.sgv}
\alias{r2.sgv}
\title{r2.sgv}
\usage{
r2.sgv(model, adj = T)
}
\arguments{
\item{model}{a fitted mixed model.}

\item{adj}{if TRUE, an adjusted R squared for model selection is provided.}
}
\value{
The R squared statistic and adjusted R squared based on SGV
        for \code{model}.
}
\description{
Computes R squared for a linear or generalized linear mixed
model using penalized quasi likelihood (PQL) estimation and standardized
generalized variance (SGV). Currently implemented for linear mixed models
with \code{\link{lmer}} and \code{\link{lme}} objects.
For generalized linear mixed models, only \code{\link{glmmPQL}}
objects are compatible.
}
\examples{

library(nlme)
m = lme(distance ~ age*Sex, random = ~1|Subject, data = Orthodont)
r2.sgv(m)

library(MASS)
PQL_mod = glmmPQL(y ~ trt + I(week > 2), random = ~ 1 | ID,
family = binomial, data = bacteria, verbose = F)
r2.sgv(PQL_mod)
}

