% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2beta_Function.R
\name{r2beta}
\alias{r2beta}
\title{Computes R^2 statistic from edwards et al., 2008.}
\usage{
r2beta(model, partial = F, ddf = "kr")
}
\arguments{
\item{model}{a fitted lmerMod object}

\item{partial}{if TRUE, semi-partial R squared are calculated for each fixed effect in the mixed model.}

\item{ddf}{Approximation method for the denominator degrees of freedom
if ddf = 'res' then residual degrees of freedom are used
if ddf = 'kr', then the Kenward Roger approach is applied.}
}
\value{
A dataframe containing the model F statistic, numerator
  and denominator degrees of freedom, and R^2 statistic. If
  partial = TRUE, then the dataframe also contains partial
  R^2 statistics for all fixed effects in the model
}
\description{
Computes R^2 statistic from edwards et al., 2008.
}
\examples{
library(nlme)
library(lme4)
m = lmer(distance ~ age*Sex+ (1|Subject), data = Orthodont)
r2beta(m, partial = T, ddf = 'kr')
}

