% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2beta_Function.R
\name{r2beta}
\alias{r2beta}
\title{r2beta}
\usage{
r2beta(model, partial = T, method = "sgv")
}
\arguments{
\item{model}{a fitted mermod, lme, or glmmPQL model.}

\item{partial}{if TRUE, semi-partial R squared are calculated for each
fixed effect in the mixed model.}

\item{method}{Specifies the method of computation for \eqn{R^{2}_{\beta}}.
if \code{method} = 'sgv' then residual degrees of freedom are used
if \code{method} = 'kr', then the Kenward Roger approach is applied.
This option is only available for \code{\link{lme}} models.
if \code{method} = 'nsj',then the Nakagawa and Schielzeth approach
is applied. This option is available for
\code{\link{lmer}} and \code{\link{lme}} objects.}
}
\value{
A dataframe containing the model F statistic, numerator
  and denominator degrees of freedom, and R squared statistic. If
  partial = TRUE, then the dataframe also contains partial
  R squared statistics for all fixed effects in the model
}
\description{
Computes coefficient of determination (R squared) from
edwards et al., 2008. Currently implemented for linear mixed models with
\code{\link{lmer}} and \code{\link{lme}} objects. For
generalized linear mixed models, \code{\link{glmmPQL}} objects
must be used with \code{ddf} = 'res'.
}
\examples{
library(nlme)
library(lme4)
linmod = lm(distance~age*Sex, data = Orthodont)
mermod = lmer(distance ~ age*Sex + (1|Subject), data = Orthodont)
lmemod = lme(distance ~ age*Sex, random = ~1|Subject, data = Orthodont)
r2beta(linmod)
r2beta(mermod, method = 'kr')
r2beta(mermod, method = 'sgv')
r2beta(mermod, method = 'nsj')


# PQL models
# Currently only residual degrees of freedom supported
library(MASS)
PQL_mod = glmmPQL(y ~ trt + I(week > 2), random = ~ 1 | ID,
                  family = binomial, data = bacteria, verbose = F)

r2beta(PQL_mod, method='sgv')
}
\references{
Edwards, Lloyd J., et al. "An R2 statistic for fixed effects in
   the linear mixed model." Statistics in medicine 27.29 (2008): 6137-6157.

Nakagawa, Shinichi, and Holger Schielzeth. "A general and simple method for
   obtaining R2 from generalized linear mixed‐effects models." Methods in
   Ecology and Evolution 4.2 (2013): 133-142.
}

